Move AnimationRunnable to its own file
