+++ app/src/main/java/com/todobom/opennotescanner/OpenNoteScannerActivity.java
@@ -10,7 +10,6 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.content.res.ColorStateList;
-import android.graphics.Bitmap;
 import android.graphics.Rect;
 import android.hardware.Camera;
 import android.media.AudioManager;
@@ -33,11 +32,6 @@ import android.view.SurfaceView;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.WindowManager;
-import android.view.animation.AccelerateInterpolator;
-import android.view.animation.Animation;
-import android.view.animation.AnimationSet;
-import android.view.animation.ScaleAnimation;
-import android.view.animation.TranslateAnimation;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.RelativeLayout;
@@ -46,6 +40,7 @@ import android.widget.Toast;
 import com.github.fafaldo.fabtoolbar.widget.FABToolbarLayout;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.navigation.NavigationView;
+import com.todobom.opennotescanner.helpers.AnimationRunnable;
 import com.todobom.opennotescanner.helpers.CustomOpenCVLoader;
 import com.todobom.opennotescanner.helpers.OpenNoteMessage;
 import com.todobom.opennotescanner.helpers.PreviewFrame;
@@ -61,7 +56,6 @@ import org.opencv.android.OpenCVLoader;
 import org.opencv.core.Core;
 import org.opencv.core.CvType;
 import org.opencv.core.Mat;
-import org.opencv.core.Point;
 import org.opencv.core.Size;
 import org.opencv.imgcodecs.Imgcodecs;
 import org.opencv.imgproc.Imgproc;
@@ -83,7 +77,6 @@ import androidx.core.content.ContextCompat;
 import androidx.fragment.app.FragmentManager;
 
 import static com.todobom.opennotescanner.helpers.Utils.addImageToGallery;
-import static com.todobom.opennotescanner.helpers.Utils.decodeSampledBitmapFromUri;
 
 /**
  * An example full-screen activity that shows and hides the system UI (i.e.
@@ -1022,130 +1015,9 @@ public class OpenNoteScannerActivity extends AppCompatActivity
         return fileName;
     }
 
-    class AnimationRunnable implements Runnable {
-
-        private Size imageSize;
-        private Point[] previewPoints = null;
-        public Size previewSize = null;
-        public String fileName = null;
-        public int width;
-        public int height;
-        private Bitmap bitmap;
-
-        public AnimationRunnable(String filename, ScannedDocument document) {
-            this.fileName = filename;
-            this.imageSize = document.processed.size();
-
-            if (document.quadrilateral != null) {
-                this.previewPoints = document.previewPoints;
-                this.previewSize = document.previewSize;
-            }
-        }
-
-        public double hipotenuse(Point a, Point b) {
-            return Math.sqrt(Math.pow(a.x - b.x, 2) + Math.pow(a.y - b.y, 2));
-        }
-
-        @Override
-        public void run() {
-            final ImageView imageView = (ImageView) findViewById(R.id.scannedAnimation);
-
-            Display display = getWindowManager().getDefaultDisplay();
-            android.graphics.Point size = new android.graphics.Point();
-            display.getRealSize(size);
-
-            int width = Math.min(size.x, size.y);
-            int height = Math.max(size.x, size.y);
-
-            // ATENTION: captured images are always in landscape, values should be swapped
-            double imageWidth = imageSize.height;
-            double imageHeight = imageSize.width;
-
-            RelativeLayout.LayoutParams params = (RelativeLayout.LayoutParams) imageView.getLayoutParams();
-
-            if (previewPoints != null) {
-                double documentLeftHeight = hipotenuse(previewPoints[0], previewPoints[1]);
-                double documentBottomWidth = hipotenuse(previewPoints[1], previewPoints[2]);
-                double documentRightHeight = hipotenuse(previewPoints[2], previewPoints[3]);
-                double documentTopWidth = hipotenuse(previewPoints[3], previewPoints[0]);
-
-                double documentWidth = Math.max(documentTopWidth, documentBottomWidth);
-                double documentHeight = Math.max(documentLeftHeight, documentRightHeight);
-
-                Log.d(TAG, "device: " + width + "x" + height + " image: " + imageWidth + "x" + imageHeight + " document: " + documentWidth + "x" + documentHeight);
-
-
-                Log.d(TAG, "previewPoints[0] x=" + previewPoints[0].x + " y=" + previewPoints[0].y);
-                Log.d(TAG, "previewPoints[1] x=" + previewPoints[1].x + " y=" + previewPoints[1].y);
-                Log.d(TAG, "previewPoints[2] x=" + previewPoints[2].x + " y=" + previewPoints[2].y);
-                Log.d(TAG, "previewPoints[3] x=" + previewPoints[3].x + " y=" + previewPoints[3].y);
-
-                // ATENTION: again, swap width and height
-                double xRatio = width / previewSize.height;
-                double yRatio = height / previewSize.width;
-
-                params.topMargin = (int) (previewPoints[3].x * yRatio);
-                params.leftMargin = (int) ((previewSize.height - previewPoints[3].y) * xRatio);
-                params.width = (int) (documentWidth * xRatio);
-                params.height = (int) (documentHeight * yRatio);
-            } else {
-                params.topMargin = height / 4;
-                params.leftMargin = width / 4;
-                params.width = width / 2;
-                params.height = height / 2;
-            }
-
-            bitmap = decodeSampledBitmapFromUri(fileName, params.width, params.height);
-
-            imageView.setImageBitmap(bitmap);
-
-            imageView.setVisibility(View.VISIBLE);
-
-            TranslateAnimation translateAnimation = new TranslateAnimation(
-                    Animation.ABSOLUTE, 0, Animation.ABSOLUTE, -params.leftMargin,
-                    Animation.ABSOLUTE, 0, Animation.ABSOLUTE, height - params.topMargin
-            );
-
-            ScaleAnimation scaleAnimation = new ScaleAnimation(1, 0, 1, 0);
-
-            AnimationSet animationSet = new AnimationSet(true);
-
-            animationSet.addAnimation(scaleAnimation);
-            animationSet.addAnimation(translateAnimation);
-
-            animationSet.setDuration(600);
-            animationSet.setInterpolator(new AccelerateInterpolator());
-
-            animationSet.setAnimationListener(new Animation.AnimationListener() {
-                @Override
-                public void onAnimationStart(Animation animation) {
-
-                }
-
-                @Override
-                public void onAnimationEnd(Animation animation) {
-                    imageView.setVisibility(View.INVISIBLE);
-                    imageView.setImageBitmap(null);
-                    if (AnimationRunnable.this.bitmap != null) {
-                        AnimationRunnable.this.bitmap.recycle();
-                    }
-                }
-
-                @Override
-                public void onAnimationRepeat(Animation animation) {
-
-                }
-            });
-
-
-            imageView.startAnimation(animationSet);
-
-        }
-    }
-
     private void animateDocument(String filename, ScannedDocument quadrilateral) {
 
-        AnimationRunnable runnable = new AnimationRunnable(filename, quadrilateral);
+        AnimationRunnable runnable = new AnimationRunnable(this, filename, quadrilateral);
         runOnUiThread(runnable);
 
     }
